Exception handling

+++ src/main/java/com/matheusmendesdev/Dscommerce/services/ProductService.java
@@ -3,10 +3,15 @@ package com.matheusmendesdev.Dscommerce.services;
 import com.matheusmendesdev.Dscommerce.dto.ProductDTO;
 import com.matheusmendesdev.Dscommerce.entities.Product;
 import com.matheusmendesdev.Dscommerce.repositories.ProductRepository;
+import com.matheusmendesdev.Dscommerce.services.exceptions.DataBaseException;
+import com.matheusmendesdev.Dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -19,7 +24,7 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id){
-        Product product = repository.findById(id).get();
+        Product product = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Resource Not Found!"));
         return new ProductDTO(product);
     }
 
@@ -39,15 +44,27 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto){
+        try{
         Product entity = repository.getReferenceById(id);
         copyDtoToEntity(entity, dto);
         entity = repository.save(entity);
         return new ProductDTO(entity);
+        }
+        catch (EntityNotFoundException ex){
+            throw new ResourceNotFoundException("Resource Not Found!");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-        repository.deleteById(id);
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException("Resouce Not Found!");
+        }
+        try{
+            repository.deleteById(id);
+        }catch (DataIntegrityViolationException ex){
+            throw new DataBaseException("Fail reference integration");
+        }
     }
 
     private void copyDtoToEntity(Product entity, ProductDTO dto) {
